# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Blossom",
        "version": "0.1"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Inscription d'utilisateur",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Le token"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Connexion d'utilisateur",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "exemple@gmail.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "password"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Le token"
                    }
                }
            }
        },
        "/api/me": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Récupération d'utilisateur",
                "operationId": "3fc3f2f762fb3c5ecf211fcb0d4f81ff",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "L'utilisateur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/plant/": {
            "get": {
                "tags": [
                    "plant"
                ],
                "summary": "Liste les plantes",
                "operationId": "3919a80ac11ad37bb04faa0d9fe9bb54",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            },
            "post": {
                "tags": [
                    "plant"
                ],
                "summary": "Ajoute une plante",
                "operationId": "8f8dd8a5063a86dbaf105f40aad62462",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "common_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "pissenlit"
                    },
                    {
                        "name": "watering_general_benchmark",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "{\"value\": \"5-7\", \"unit\": \"days\"}"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "La plante a bien été ajouté"
                    }
                }
            }
        },
        "/api/plant/{name}": {
            "get": {
                "tags": [
                    "plant"
                ],
                "summary": "Montre une plante à partir de son nom",
                "operationId": "7689b9b14a12aa70fcbe135e7f73bc32",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "rose"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réussite",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Réussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/plant/{id}": {
            "delete": {
                "tags": [
                    "plant"
                ],
                "summary": "Supprime une plante",
                "operationId": "533f3fb60e61315d56125d3e1efeb83b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réussite",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Réussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Plante non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Non trouvée"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/plant/": {
            "post": {
                "tags": [
                    "user_plant"
                ],
                "summary": "Ajoute une plante à un utilisateur",
                "description": "Show the form for creating a new resource.",
                "operationId": "07fdd7b895abc5129c3fcf8104b346d4",
                "requestBody": {
                    "description": "Json payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "city",
                                    "country"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "rose"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Roanne"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "France"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Réussite",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Réussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/plant/{id}": {
            "delete": {
                "tags": [
                    "user_plant"
                ],
                "summary": "Supprime une plante d'utilisateur",
                "description": "Remove the specified resource from storage.",
                "operationId": "ee358d3a567b9f3c3986f4d2686254bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de la plante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réussite",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Réussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "user",
            "description": "user"
        },
        {
            "name": "plant",
            "description": "plant"
        },
        {
            "name": "user_plant",
            "description": "user_plant"
        }
    ]
}